pyproject.toml
[project]
name = "british-cycling-utils"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "elliot-100", email = "3186037+elliot-100@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "mypy>=1.16.1",
    "pydantic>=2.11.7",
    "python-dateutil>=2.9.0",
    "types-python-dateutil>=2.9.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.12.1",
]
